Move variable to the ProviderContext
